Update modules and styles

+++ .config/waybar/config_watch.sh
@@ -1,10 +1,19 @@
 #!/bin/sh
 
 # while read -r directory events filename; do
-inotifywait -e close_write,moved_to,create -m . |
+inotifywait -e close_write -m . |
 while read -r _ events filename; do
+    [ "$events" != "CLOSE_WRITE,CLOSE" ] && continue
+
+    echo "before" "$events" "$filename"
+
+    if [ "$filename" != "config.jsonc" ] && [ "$filename" != "style.css" ]; then 
+        continue
+    fi
+
     echo "$events $filename"
+
     killall waybar
-    hyprctl dispatch exec -- GTK_DEBUG=interactive waybar
+    hyprctl dispatch exec -- waybar
 done
